all cruds tested

+++ src/categroies/categroies.controller.ts
@@ -6,12 +6,14 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { CategroiesService } from './categroies.service';
 import { CreateCategroyDto } from './dto/create-categroy.dto';
 import { UpdateCategroyDto } from './dto/update-categroy.dto';
+import { isUUID } from 'class-validator';
 
-@Controller('categroies')
+@Controller('categories')
 export class CategroiesController {
   constructor(private readonly categroiesService: CategroiesService) {}
 
@@ -25,8 +27,16 @@ export class CategroiesController {
     return this.categroiesService.findAll();
   }
 
+  @Get('/paginated')
+  findWithPaginationAll() {
+    return this.categroiesService.findCategoriesWithPaginatedProducts();
+  }
+
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.categroiesService.findOne(id);
   }
 
@@ -35,11 +45,17 @@ export class CategroiesController {
     @Param('id') id: string,
     @Body() updateCategroyDto: UpdateCategroyDto,
   ) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.categroiesService.update(id, updateCategroyDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.categroiesService.remove(id);
   }
 }
