+++ src/order-item/order-item.controller.ts
@@ -6,10 +6,12 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { OrderItemService } from './order-item.service';
 import { CreateOrderItemDto } from './dto/create-order-item.dto';
 import { UpdateOrderItemDto } from './dto/update-order-item.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('order-item')
 export class OrderItemController {
@@ -35,11 +37,17 @@ export class OrderItemController {
     @Param('id') id: string,
     @Body() updateOrderItemDto: UpdateOrderItemDto,
   ) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.orderItemService.update(id, updateOrderItemDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.orderItemService.remove(id);
   }
 }
