+++ src/products/products.controller.ts
@@ -6,10 +6,12 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('products')
 export class ProductsController {
@@ -27,16 +29,25 @@ export class ProductsController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.productsService.findOne(id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.productsService.update(id, updateProductDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.productsService.remove(id);
   }
 }
