+++ src/reviews/reviews.controller.ts
@@ -6,10 +6,12 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { ReviewsService } from './reviews.service';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('reviews')
 export class ReviewsController {
@@ -27,16 +29,25 @@ export class ReviewsController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.reviewsService.findOne(id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateReviewDto: UpdateReviewDto) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.reviewsService.update(id, updateReviewDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID format');
+    }
     return this.reviewsService.remove(id);
   }
 }
